fix(users): accepte un tableau de roles dans l'annuaire (TypeError 500)
Les selects multi-roles du front (affectation mission, page Statistiques)
envoient role[]=admin&role[]=collaborateur ; UserService::listerAnnuaire
etait type ?string -> TypeError -> 500 -> toast d'erreur a l'ouverture de
la page Statistiques (et du select d'affectation de MissionDetailPage).
Signature elargie a array|string|null (le scope Spatie role() accepte
nativement les tableaux) + 2 tests de non-regression (tableau + unique).

Au passage : filtre du delai moyen de tache type float (Carbon 3 renvoie
des diffInDays decimaux -> deprecation de conversion implicite).

+++ backend/app/Services/KpiService.php
@@ -206,8 +206,9 @@ class KpiService
                 ->when($exerciceId, fn ($q) => $q->where('exercice_id', $exerciceId))
             )
             ->get(['created_at', 'updated_at'])
+            // Carbon 3 : diffInDays renvoie un float (jours decimaux) — ne pas typer int
             ->map(fn (Tache $t) => $t->created_at->diffInDays($t->updated_at))
-            ->filter(fn (int $d) => $d >= 0);
+            ->filter(fn (float $d) => $d >= 0);
 
         if ($delais->isEmpty()) {
             return 0.0;
